Eliminate lodash.isequal dependency
