+++ src/style-spec/diff.js
@@ -1,5 +1,5 @@
 
-const isEqual = require('lodash.isequal');
+const isEqual = require('./util/deep_equal');
 
 const operations = {
