+++ src/style-spec/package.json
@@ -50,7 +50,6 @@
     "csscolorparser": "~1.0.2",
     "in-publish": "^2.0.0",
     "jsonlint-lines-primitives": "~1.6.0",
-    "lodash.isequal": "^3.0.4",
     "minimist": "0.0.8",
     "rw": "^1.3.3",
     "sort-object": "^0.3.2",
