+++ src/util/util.js
@@ -288,30 +288,7 @@ exports.filterObject = function(input: Object, iterator: Function, context?: Obj
     return output;
 };
 
-/**
- * Deeply compares two object literals.
- *
- * @private
- */
-exports.deepEqual = function(a: ?mixed, b: ?mixed): boolean {
-    if (Array.isArray(a)) {
-        if (!Array.isArray(b) || a.length !== b.length) return false;
-        for (let i = 0; i < a.length; i++) {
-            if (!exports.deepEqual(a[i], b[i])) return false;
-        }
-        return true;
-    }
-    if (typeof a === 'object' && a !== null && b !== null) {
-        if (!(typeof b === 'object')) return false;
-        const keys = Object.keys(a);
-        if (keys.length !== Object.keys(b).length) return false;
-        for (const key in a) {
-            if (!exports.deepEqual(a[key], b[key])) return false;
-        }
-        return true;
-    }
-    return a === b;
-};
+exports.deepEqual = require('../style-spec/util/deep_equal');
 
 /**
  * Deeply clones two objects.
